# backend/api/workflow/v1/workflow_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             v3.21.12
// source: workflow/v1/workflow.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationWorkflowExecuteWorkflow = "/workflow.v1.Workflow/ExecuteWorkflow"
const OperationWorkflowGetWorkflowStatus = "/workflow.v1.Workflow/GetWorkflowStatus"
const OperationWorkflowListWorkflowExecutions = "/workflow.v1.Workflow/ListWorkflowExecutions"

type WorkflowHTTPServer interface {
	// ExecuteWorkflow Execute a workflow
	ExecuteWorkflow(context.Context, *ExecuteWorkflowRequest) (*ExecuteWorkflowReply, error)
	// GetWorkflowStatus Get workflow execution status
	GetWorkflowStatus(context.Context, *GetWorkflowStatusRequest) (*GetWorkflowStatusReply, error)
	// ListWorkflowExecutions List workflow executions
	ListWorkflowExecutions(context.Context, *ListWorkflowExecutionsRequest) (*ListWorkflowExecutionsReply, error)
}

func RegisterWorkflowHTTPServer(s *http.Server, srv WorkflowHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/workflow/execute", _Workflow_ExecuteWorkflow0_HTTP_Handler(srv))
	r.GET("/v1/workflow/status/{execution_id}", _Workflow_GetWorkflowStatus0_HTTP_Handler(srv))
	r.GET("/v1/workflow/executions", _Workflow_ListWorkflowExecutions0_HTTP_Handler(srv))
}

func _Workflow_ExecuteWorkflow0_HTTP_Handler(srv WorkflowHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ExecuteWorkflowRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationWorkflowExecuteWorkflow)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ExecuteWorkflow(ctx, req.(*ExecuteWorkflowRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ExecuteWorkflowReply)
		return ctx.Result(200, reply)
	}
}

func _Workflow_GetWorkflowStatus0_HTTP_Handler(srv WorkflowHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetWorkflowStatusRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationWorkflowGetWorkflowStatus)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetWorkflowStatus(ctx, req.(*GetWorkflowStatusRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetWorkflowStatusReply)
		return ctx.Result(200, reply)
	}
}

func _Workflow_ListWorkflowExecutions0_HTTP_Handler(srv WorkflowHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListWorkflowExecutionsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationWorkflowListWorkflowExecutions)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListWorkflowExecutions(ctx, req.(*ListWorkflowExecutionsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListWorkflowExecutionsReply)
		return ctx.Result(200, reply)
	}
}

type WorkflowHTTPClient interface {
	// ExecuteWorkflow Execute a workflow
	ExecuteWorkflow(ctx context.Context, req *ExecuteWorkflowRequest, opts ...http.CallOption) (rsp *ExecuteWorkflowReply, err error)
	// GetWorkflowStatus Get workflow execution status
	GetWorkflowStatus(ctx context.Context, req *GetWorkflowStatusRequest, opts ...http.CallOption) (rsp *GetWorkflowStatusReply, err error)
	// ListWorkflowExecutions List workflow executions
	ListWorkflowExecutions(ctx context.Context, req *ListWorkflowExecutionsRequest, opts ...http.CallOption) (rsp *ListWorkflowExecutionsReply, err error)
}

type WorkflowHTTPClientImpl struct {
	cc *http.Client
}

func NewWorkflowHTTPClient(client *http.Client) WorkflowHTTPClient {
	return &WorkflowHTTPClientImpl{client}
}

// ExecuteWorkflow Execute a workflow
func (c *WorkflowHTTPClientImpl) ExecuteWorkflow(ctx context.Context, in *ExecuteWorkflowRequest, opts ...http.CallOption) (*ExecuteWorkflowReply, error) {
	var out ExecuteWorkflowReply
	pattern := "/v1/workflow/execute"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationWorkflowExecuteWorkflow))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetWorkflowStatus Get workflow execution status
func (c *WorkflowHTTPClientImpl) GetWorkflowStatus(ctx context.Context, in *GetWorkflowStatusRequest, opts ...http.CallOption) (*GetWorkflowStatusReply, error) {
	var out GetWorkflowStatusReply
	pattern := "/v1/workflow/status/{execution_id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationWorkflowGetWorkflowStatus))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListWorkflowExecutions List workflow executions
func (c *WorkflowHTTPClientImpl) ListWorkflowExecutions(ctx context.Context, in *ListWorkflowExecutionsRequest, opts ...http.CallOption) (*ListWorkflowExecutionsReply, error) {
	var out ListWorkflowExecutionsReply
	pattern := "/v1/workflow/executions"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationWorkflowListWorkflowExecutions))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}