# backend/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/workflow/execute:
        post:
            tags:
                - Workflow
            description: Execute a workflow
            operationId: Workflow_ExecuteWorkflow
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/workflow.v1.ExecuteWorkflowRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/workflow.v1.ExecuteWorkflowReply'
    /v1/workflow/executions:
        get:
            tags:
                - Workflow
            description: List workflow executions
            operationId: Workflow_ListWorkflowExecutions
            parameters:
                - name: userId
                  in: query
                  schema:
                    type: string
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/workflow.v1.ListWorkflowExecutionsReply'
    /v1/workflow/status/{executionId}:
        get:
            tags:
                - Workflow
            description: Get workflow execution status
            operationId: Workflow_GetWorkflowStatus
            parameters:
                - name: executionId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/workflow.v1.GetWorkflowStatusReply'
components:
    schemas:
...
        workflow.v1.ExecuteWorkflowReply:
            type: object
            properties:
                executionId:
                    type: string
                status:
                    type: string
                message:
                    type: string
            description: Reply for executing a workflow
        workflow.v1.ExecuteWorkflowRequest:
            type: object
            properties:
                workflowType:
                    type: string
                parameters:
                    type: object
                    additionalProperties:
                        type: string
                userId:
                    type: string
            description: Request for executing a workflow
        workflow.v1.GetWorkflowStatusReply:
            type: object
            properties:
                executionId:
                    type: string
                status:
                    type: string
                result:
                    type: string
                errorMessage:
                    type: string
                createdAt:
                    type: string
                updatedAt:
                    type: string
            description: Reply for getting workflow status
        workflow.v1.ListWorkflowExecutionsReply:
            type: object
            properties:
                executions:
                    type: array
                    items:
                        $ref: '#/components/schemas/workflow.v1.WorkflowExecution'
                total:
                    type: integer
                    format: int32
            description: Reply for listing workflow executions
        workflow.v1.WorkflowExecution:
            type: object
            properties:
                id:
                    type: string
                workflowType:
                    type: string
                status:
                    type: string
                userId:
                    type: string
                createdAt:
                    type: string
                updatedAt:
                    type: string
            description: Workflow execution structure
...